backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
motor==3.6.0
pymongo>=4.9.0,<4.10
beanie==1.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
python-dotenv==1.0.1
faker==27.0.0
email-validator==2.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
